style: QuoteCard components.
